Able to get fetch slicer and printer profile data and display on SlicingFragment.

+++ data/src/main/java/com/nairbspace/octoandroid/data/entity/SlicingProgressEntity.java
@@ -14,5 +14,5 @@ public abstract class SlicingProgressEntity {
     @Nullable @SerializedName("source_patch") public abstract String sourcePath();
     @Nullable @SerializedName("dest_location") public abstract String destLocation();
     @Nullable @SerializedName("dest_path") public abstract String destPath();
-    @Nullable @SerializedName("progress") public abstract Float progress();
+    @Nullable @SerializedName("progress") public abstract Double progress();
 }
